requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.1
gunicorn>=21.2.0
Pillow>=10.0
django-cors-headers>=4.7.0